requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.3.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
polib==1.2.0